airflow/requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-google
apache-airflow-providers-airbyte
dbt-postgres==1.6.6
dbt-bigquery==1.6.7
astronomer-cosmos==1.1.3